requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
tenacity
sqlglot
pymysql
psycopg[binary]
redis
PyYAML
tqdm
cryptography
cachetools>=5.3.3